Fix bug where widening could fail randomly.
The previous logic of reducing elements in domains.Set was used to
join or update two sets by reducing their concatenation. By merging
this way, we were losing the ordering information, which was
problematic because "update" is in general not commutative, thus
causing widening to fail when the (random) order was not the one
expected.

Adds a method "_merge" that is aware of the initial ordering.

Change-Id: Ibc94bea8f59b474c4b00bd91203fc695c6df8910