Modify how CFG node's data is set by CFGBuilder.
A CFG node can have two attributes for now: 'is_widening_point' or
'node'. Previously, these were set only for nodes for which it was
relevant.

With this commit, these attributes are set for all nodes but their value
indicates if it is relevant for the corresponding CFG node.

Allows to remove 'hasattr' usage.

Change-Id: I7886686d0d7d1989d6b3e08e6c85cbe2b7522801