Add TypeInterpreter and some default interpretations.
TypeInterpreter is the tool that determines how a type should be interpreted
by the checkers. Concretely, for a given type, it returns the domain used
to represent it as well as a dictionary of definitions that are available
for this new domain.

A default interpretation for the Boolean type and the IntRange type are
provided.

Change-Id: Ie27bdbc936881a0710347744cdbe32a85f05cc48