Insert synthetic assume statement before pointer dereference.
When a pointer ptr is dereferenced in an expression, an
"assume(ptr != null)" statement is inserted right before it.

This is done on order to emulate the fact that only a non-null
pointer can be dereferenced. The assume statement has a flag
that can be used by checkers to detect null dereferences.

Change-Id: Ie05c314b480a48a7acaf624b6cdf084c3b4edee7