Convert universal integer types to the appropriate type.
This change is needed because of the recent change in
libadalang to type int/real literals by their universal
type.

It tries to perform the implicit conversion, by using
the expected type of the result, or the type of one
of the other operand if it is not also an universal type.

Change-Id: I82a8ed1fddf5db8c2336a3ea25ede81640dbd3cf