Update python driver, adds support for the "helper" option.
When this option is set, the driver will not run a python on
a "test.py" file, but on the specified test helper located in
"tests/test_helpers/". It will also invoke it with a set of
arguments. (For now just 'output_dir', which can be used by the
test to output persistent data such as the graphs generated
by checkers.)

Change-Id: Id2a6760e8471e5f7a2c7ba6224a51efc2540da76