QB28-022: Modify logic for representing unique variables in Basic IR.
Previously, a variable in the Basic IR tree was represented by
the same identifier at each of its use site for simplicity. This
made it difficult to retrieve the original position in the source
code of these identifiers. Now, a new Identifier is used at each
use site of a variable. However, a unique Variable node (new
addition) is accessible from each of these Identifiers.
Overall, only a level of indirection was added.

Change-Id: I97b5a160ba54bdc750236c1739416944476025ee