Fix bug in windows where logger could be unset in child processes.
Again, due to global variables not being copied to child processes
under windows, the global logger was None in child processes and
therefore the messages produced by the child processes were never
sent to any logger. The fix is to simply reset the logger at the
start of each child process.

Change-Id: I9912590a557b3a75a2049a0c179616d57c48c9be