Fix undeterminism in testsuite output due to recent changes.
Since the addition of the pre-analysis done before IR generation,
subprograms are not necessarily transformed in the order they
appear inside the code. Therefore, the tests were sometimes
performed on the wrong subprograms since the test scripts worked
on the first subprogram returned by the IR generation routine.

The fix consists in naming explictly the functions/procedures
to test in the different test cases and using the "test_subprogram"
option in the test YAML. The different test_helpers now look
specifically for the subprogram that has required name to run on.
If no "test_subprogram" is specified, "Test" is used.

Change-Id: Ic7b03036cf9482f045c28d92746585203182381e