Make the model data structure indexed by subprograms.
Previously, a model contained information about all the nodes
appearing in all subprograms that were given to the model creation
routine. After this change, the model is first indexed by
subprogram, and then by node. This means that when analyzing a
given subprogram, the model is not polluted with information about
other subprograms.

Change-Id: Iea1c3c8db00eb5a09832fe6ea2bc9eb7552ea166