Fix checkers' logic for evaluating an assume expression.
Previously, the representation of an element of the booleans domain
had to be assumed in each checker as being a subset of {"True", "False"}.
With this change, the checkers don't make any assumption anymore. Instead,
they concretize the abstract value and check if True and False belong to
the set of concrete values or not. In parallel, the default Boolean domain
was modified to ease manipulations of its elements within the checkers.

Change-Id: Ia3c6b3c5eb5d0f038512c656a0ab34e5d8ffef29