RB02-013: Force shutdown of manager object and ignore WindowsErrors.
For some reasons, termination of the server process started by the Manager
object fails with a WindowsError. By default, this process is being
terminated in a finalizer at the end of the program, meaning user code
cannot catch that error. This patch forces the shutdown of the Manager
after parallel_loop has finished its execution and catches (ignores) any
related WindowsError.

Change-Id: If4ba055b1ee7e73914d38f7e4b7476cf714b0518