SC03-033: Compute widening points explicitly.
Until now we would rely on the "Loop" IR construct to infer
widening points. The problem is that it is possible to generate
loops using "Goto" statements, and we would not infer any
widening point for such loops.

This change fixes this by computing widening points during
a depth-first search on the CFG, by identifying back edges.

I took the opportunity to mix this analysis with the existing
reachability analysis so that we do only one DFS.

Change-Id: If5b82ce5ac293e32f263403d13bd6d3d9ffb4f99